# api/openapi/notifications.yml
openapi: 3.0.3
info:
  title: Notifications API
  description: API for managing in-app notifications
  version: 1.0.0
servers:
  - url: /api
paths:
  /notifications:
    post:
      summary: Create a notification (for testing/admin)
      operationId: createNotification
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - message
                - type
              properties:
                title:
                  type: string
                message:
                  type: string
                type:
                  type: string
                  enum: [streak_milestone, habit_reminder, achievement, system, welcome]
                data:
                  type: object
      responses:
        '201':
          description: Notification created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
      security:
        - bearerAuth: []
    get:
      summary: List user notifications
      operationId: listNotifications
      description: Get a paginated list of notifications for the current user
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: unread_only
          in: query
          description: Only return unread notifications
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: A paginated list of notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotificationListWrapper'
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: object
      properties:
        has_previous_page:
          type: boolean
        has_next_page:
          type: boolean
        current_page:
          type: integer
        per_page:
          type: integer
        total_data:
          type: integer
        total_data_in_current_page:
          type: integer
        last_page:
          type: integer
        from:
          type: integer
        to:
          type: integer
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [streak_milestone, habit_reminder, achievement, system, welcome]
        title:
          type: string
        message:
          type: string
        data:
          type: object
          additionalProperties: true
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
    PaginatedNotificationListWrapper:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...